backend/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
openai>=1.0.0
pydantic>=1.8.0

# Server and ASGI
gunicorn>=20.1.0

# Optional but recommended
python-multipart>=0.0.5  # For handling form data
ujson>=4.0.0  # Faster JSON processing
httptools>=0.3.0  # Faster HTTP processing
uvloop>=0.16.0  # Faster event loop

# Testing and development (optional)
pytest>=6.2.5
requests>=2.26.0